Get page Url from _SERVER['SCRIPT_URL']
_GET['q'] was being used to store a profiled page's url, however
it is not accurate for my site.  For example, I would like the Url
for a request to http://mysite.com/ to be stored as '/', but
_GET['q'] contained the value 'reviews'.  Not sure if this is site-
specific, or a more general Drupal issue.  Using _SERVER['SCRIPT_URL']
fixes the problem for me.

Perhaps this should be configurable in admin/settings/xhprof ??