Don't store profile blob in the database.  Store only to filesystem.
At my site we don't want to store profile data blobs to the database.
This causes unwanted write activity on the production database master.
Instead, we configure the xhprof module in php.ini to write to a shared
filesystem mounted by all webservers in the cluster.

I realize that instead of hard-coding this, it should be configurable
via the XHProf Settings page (admin/settings/xhprof).
Feature request anyone?